refactory slightly and remove console.logs

+++ controllers/employeesController.js
@@ -2,25 +2,31 @@ const Employee = require("../model/Employee");
 
 const getAllEmployees = async (req, res) => {
   const employees = await Employee.find({});
+  if (!employees) {
+    return res.status(204).json({ message: "No employees found" });
+  }
   res.json(employees);
 };
 
 const createNewEmployee = async (req, res) => {
-  const { firstname, lastname } = req.body;
-  if (!firstname || !lastname) {
+  if (!req.body?.firstname || !req.body?.lastname) {
     return res
       .status(400)
       .json({ message: "First and last names are required" });
   }
+  const { firstname, lastname } = req.body;
   const employee = await Employee.create({ firstname, lastname });
   res.status(201).json(employee);
 };
 
 const updateEmployee = async (req, res) => {
+  if (!req.body?.id) {
+    return res.status(400).json({ message: "id is required" });
+  }
   const employee = await Employee.findById(req.body.id);
   if (!employee) {
     return res
-      .status(400)
+      .status(404)
       .json({ message: `Employee ID ${req.body.id} not found` });
   }
   if (req.body.firstname) employee.firstname = req.body.firstname;
@@ -30,17 +36,23 @@ const updateEmployee = async (req, res) => {
 };
 
 const deleteEmployee = async (req, res) => {
+  if (!req.body?.id) {
+    return res.status(400).json({ message: "id is required" });
+  }
   const employee = await Employee.findByIdAndDelete(req.body.id);
   if (!employee) {
     return res
       .status(400)
       .json({ message: `Employee ID ${req.body.id} not found` });
   }
-  res.sendStatus(204);
+  res.json(employee);
 };
 
-const getEmployee = (req, res) => {
-  const employee = Employee.findById(req.params.id);
+const getEmployee = async (req, res) => {
+  if (!req.params?.id) {
+    return res.status(400).json({ message: "id parameter is required" });
+  }
+  const employee = await Employee.findById(req.params.id);
   if (!employee) {
     return res
       .status(400)
